Rename index-domready to index, index to iife
Rename the IIFE blocking bundle from index.ts to iife.ts and rename
the module entry from index-domready.ts to index.ts. This gives the
main CSS output a clean name (css/index.css) instead of the previous
css/index-domready.css.
