+++ templates/base/head_style.tmpl
@@ -1,2 +1,2 @@
-<link rel="stylesheet" href="{{AssetUrlPrefix}}/{{GetAssetPath "css/index-domready.css"}}">
+<link rel="stylesheet" href="{{AssetUrlPrefix}}/{{GetAssetPath "css/index.css"}}">
 <link rel="stylesheet" href="{{AssetUrlPrefix}}/{{GetAssetPath (printf "css/theme-%s.css" (PathEscape ctx.CurrentWebTheme.InternalName))}}">
